TSDK-566 First working version of the broker.
