--- daml-topl-broker/src/main/scala/Main.scala
@@ -1,6 +1,0 @@
-
-object Main {
-  println("Hello World!")
-
-}
-
