+++ daml-topl-broker/src/main/scala/co/topl/broker/BrokerCLIParamConfig.scala
@@ -1,59 +1,22 @@
-import scopt.OParser
-import co.topl.brambl.constants.NetworkConstants
+package co.topl.broker
 
-case class CLIParamConfig(
+import co.topl.shared.NetworkIdentifiers
+import co.topl.shared.InvalidNet
+import scopt.OParser
+
+case class BrokerCLIParamConfig(
   damlHost:            String = "localhost",
   damlPort:            Int = 6865,
   damlHub:             Boolean = false,
   damlSecurityEnabled: Boolean = false,
   network:             NetworkIdentifiers = InvalidNet,
   damlAccessToken:     Option[String] = None,
-  dappParty:           String = "",
-  operatorParty:       String = "",
-  password:            String = ""
+  operatorParty:       String = ""
 )
 
-sealed abstract class NetworkIdentifiers(
-  val i:         Int,
-  val name:      String,
-  val networkId: Int
-) {
-  override def toString: String = name
-}
+private [broker] trait ParameterProcessorModule {
 
-case object NetworkIdentifiers {
-
-  def values = Set(Mainnet, Testnet, Privatenet)
-
-  def fromString(s: String): Option[NetworkIdentifiers] =
-    s match {
-      case "mainnet" => Some(Mainnet)
-      case "testnet" => Some(Testnet)
-      case "private" => Some(Privatenet)
-      case _         => None
-    }
-}
-
-case object Mainnet extends NetworkIdentifiers(0, "mainnet", NetworkConstants.MAIN_NETWORK_ID)
-case object Testnet extends NetworkIdentifiers(1, "testnet", NetworkConstants.TEST_NETWORK_ID)
-
-case object Privatenet
-    extends NetworkIdentifiers(
-      2,
-      "private",
-      NetworkConstants.PRIVATE_NETWORK_ID
-    )
-
-case object InvalidNet
-    extends NetworkIdentifiers(
-      -1,
-      "invalid",
-      NetworkConstants.PRIVATE_NETWORK_ID
-    )
-
-trait ParameterProcessorModule {
-
-  val builder = OParser.builder[CLIParamConfig]
+  val builder = OParser.builder[BrokerCLIParamConfig]
 
   implicit val networkRead: scopt.Read[NetworkIdentifiers] =
     scopt.Read
@@ -69,7 +32,7 @@ trait ParameterProcessorModule {
   val parser = {
     import builder._
     OParser.sequence(
-      programName("daml-topl-dapp"),
+      programName("daml-topl-broker"),
       head("daml-topl-dapp", "0.1"),
       opt[String]("daml-host")
         .action((x, c) => c.copy(damlHost = x))
@@ -83,14 +46,6 @@ trait ParameterProcessorModule {
       opt[Boolean]('s', "daml-security-enabled")
         .action((x, c) => c.copy(damlSecurityEnabled = x))
         .text("whether to use TLS for the connection to the ledger"),
-      opt[String]("dapp-party")
-        .action((x, c) => c.copy(dappParty = x))
-        .text("the party that will be used to submit transactions to the ledger")
-        .validate(x =>
-          if (x.isEmpty) failure("dapp-party must not be empty")
-          else success
-        )
-        .required(),
       opt[NetworkIdentifiers]('n', "network")
         .action((x, c) => c.copy(network = x))
         .text(
@@ -109,14 +64,6 @@ trait ParameterProcessorModule {
       opt[Option[String]]('t', "daml-access-token")
         .action((x, c) => c.copy(damlAccessToken = x))
         .text("the access token for the ledger"),
-      opt[String]('w', "password")
-        .action((x, c) => c.copy(password = x))
-        .text("the password for the keyfile")
-        .validate(x =>
-          if (x.isEmpty) failure("password must not be empty")
-          else success
-        )
-        .required()
     )
   }
