+++ daml-topl-broker/src/main/scala/co/topl/broker/processors/DAMLAlgebra.scala
@@ -1,4 +1,4 @@
-package processors
+package co.topl.broker.processors
 
 import java.util.UUID
 import java.util.stream
