+++ daml-topl-broker/src/main/scala/co/topl/broker/processors/SimpleTransactionAlgebra.scala
